blog post delete feature

+++ lib/fitness_blogs/blocs/blog_post_delete/blog_post_delete_bloc.dart
@@ -10,9 +10,10 @@ part 'blog_post_delete_state.dart';
 class BlogPostDeleteBloc extends Bloc<BlogPostDeleteEvent, BlogPostDeleteState> {
   BlogPostDeleteBloc({
     required String blogId,
+    required bool isAuthor,
     required BlogRepository blogRepository,
   })  : _blogRepository = blogRepository,
-        super(BlogPostDeleteInitial(blogId));
+        super(BlogPostDeleteInitial(blogId, isAuthor));
 
   final BlogRepository _blogRepository;
 
@@ -26,14 +27,18 @@ class BlogPostDeleteBloc extends Bloc<BlogPostDeleteEvent, BlogPostDeleteState> 
   }
 
   Stream<BlogPostDeleteState> _mapDeleteRequestedToState(BlogPostDeleteRequested event) async* {
-    yield BlogPostDeleteLoading(state.blogId);
+    yield BlogPostDeleteLoading(state.blogId, state.isAuthor);
 
     try {
-      await _blogRepository.deleteBlogPost(state.blogId);
+      if (state.isAuthor) {
+        await _blogRepository.deleteBlogPost(state.blogId);
 
-      yield BlogPostDeleteSuccess(state.blogId);
+        yield BlogPostDeleteSuccess(state.blogId, state.isAuthor);
+      } else {
+        yield BlogPostDeleteFail(state.blogId, state.isAuthor);
+      }
     } catch (error) {
-      yield BlogPostDeleteFail(state.blogId);
+      yield BlogPostDeleteFail(state.blogId, state.isAuthor);
     }
   }
 }
